third_party/blink/web_tests/fast/dom/shadow/suppress-mutation-events-in-shadow-characterdata-expected.txt

Ensures that DOMCharacterDataModified isn't fired inside shadows.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fired is false
PASS div.innerHTML is 'Hello, World!'
PASS successfullyParsed is true

TEST COMPLETE
